Support pytorch dataframes
