+++ examples/sin.py
@@ -1,26 +1,28 @@
 import pandas as pd
 import numpy as np
 from matplotlib import pyplot as plt
-
+import torch
 from bartpy.initializers.initializer import Initializer
 from bartpy.sklearnmodel import SklearnModel
-from bartpy.diagnostics.trees import plot_tree_depth
-from bartpy.diagnostics.features import plot_feature_split_proportions
-from bartpy.diagnostics.residuals import plot_qq
-
-from bartpy.samplers.oblivioustrees.treemutation import get_tree_sampler
-import statsmodels.api as sm
-from bartpy.extensions.baseestimator import ResidualBART
-from bartpy.extensions.ols import OLS
-
+# from bartpy.diagnostics.trees import plot_tree_depth
+# from bartpy.diagnostics.features import plot_feature_split_proportions
+# from bartpy.diagnostics.residuals import plot_qq
+#
+# from bartpy.samplers.oblivioustrees.treemutation import get_tree_sampler
+# import statsmodels.api as sm
+# from bartpy.extensions.baseestimator import ResidualBART
+# from bartpy.extensions.ols import OLS
+#
 
 def run(alpha, beta, n_trees, size=100):
     import warnings
 
     warnings.simplefilter("error", UserWarning)
     x = np.linspace(0, 5, size)
-    X = pd.DataFrame(x)
     y = np.random.normal(0, 1.0, size=size) + np.sin(x)
+    X = pd.DataFrame(x)
+    # X = torch.from_numpy(x.reshape(-1, 1))
+    # y = torch.from_numpy(y)
     from bartpy.samplers.unconstrainedtree.treemutation import get_tree_sampler
 
     model = SklearnModel(
@@ -35,7 +37,7 @@ def run(alpha, beta, n_trees, size=100):
     plt.plot(y)
     plt.plot(model.predict(X))
     plt.show()
-    # plot_tree_depth(model)
+    # # plot_tree_depth(model)
     # plot_feature_split_proportions(model)
     # plot_qq(model)
     # null_distr = null_feature_split_proportions_distribution(model, X, y)
@@ -48,6 +50,6 @@ if __name__ == "__main__":
     from datetime import datetime as dt
     print(dt.now())
 
-    run(0.95, 2., 200, size=500)
-    #cProfile.run("run(0.95, 2., 200, size=500)", "restats")
+    #run(0.95, 2., 200, size=500000)
+    cProfile.run("run(0.95, 2., 200, size=500)", "restatsto")
     print(dt.now())
